[ops, test] fix: keep eager MoE compute dtype under fp32 routing
Pin the SwiGLU routing scale to the activation dtype so batch-invariant
linear does not mix fp32 router scores with bf16 expert weights. Align
the Wan padded Ulysses reference to the same mask, skip Qwen-Image
isolation on NPU CPU tensors, and give DeepSeek-V3 e2e the same bf16
mixed-precision extra_args as V4.

+++ veomni/ops/kernels/moe_experts/standard/eager.py
@@ -62,8 +62,10 @@ def wrapper(
             gate = gate.clamp(max=swiglu_limit)
             up = up.clamp(min=-swiglu_limit, max=swiglu_limit)
         y = F.silu(gate) * up
-        # Fused order: scale the intermediate, then fc2.
-        y = y * routing_weights[token_idx, top_k_pos, None]
+        # Fused order: scale the intermediate, then fc2. Router scores stay
+        # float32 even under FSDP2 bf16 compute; keep the scale on ``y.dtype``
+        # so batch-invariant ``F.linear`` sees matching activation/weight dtypes.
+        y = y * routing_weights[token_idx, top_k_pos, None].to(dtype=y.dtype)
         y = F.linear(y, fc2_weight[idx])
         output = output.index_add(0, token_idx, y.to(output.dtype))
     return output
